Create links-citations
This is the data associated with the connection between a given pair of attackers: the connection itself, and eventually the material demonstrating this connection (eg clicking the visible line representing the link in the network will take you to someone's manifesto referencing the other person). Right now I have placeholder links in play.