Add files via upload
All current edges with hyperlink to original documents and quoted reference. This may not be the ONLY reference in the entire document. Due to the technicalities of my OCR, I kept getting duplicates and it was hard to tell what was novel, so (being quite young technically at the time) I made the decision to get one good one per connection and leave the rest out.